test/채팅 단위 및 통합테스트의 케이스 분리 및 커버리지 작업 #41
